pyproject.toml
[tool.poetry]
name = "justopinion"
version = "0.1.0"
description = "Download client for judicial opinions"
authors = ["Matt Carey <matt@authorityspoke.com>"]
license = "Atmosphere Software License Version 0.4–🚪🛂"

[tool.poetry.dependencies]
python = "^3.9"
anchorpoint = { git = "https://github.com/mscarey/anchorpoint.git", rev = "72e3154" }
eyecite = "^2.2.0"
requests = "^2.25.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "3.9.2"
mypy = "^0.910"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-recording = "^0.11.0"
python-dotenv ="^0.18.0"
Sphinx = "^4.0.3"
sphinx-autodoc-typehints = "^1.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
